Estilizando tela detail Adicionado Background Image
