+++ pokedex/src/Routes/Router.js
@@ -1,6 +1,6 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { Home } from "../pages/Home.js";
-import { DetailPage } from "../pages/DetailPage.js";
+import { DetailPage } from "../pages/DetailPage/DetailPage.js";
 import { Pokedex } from "../pages/Pokedex.js";
 import { Header } from "../Components/Header/Header.js";
